requirements.txt
flask==2.0.1
werkzeug==2.0.3
pywebview==3.7.2
flask-cors==3.0.10
pillow==9.5.0
pyinstaller==5.13.0
openai
easyocr
pystray
pywin32
numpy